+++ packages/base-zone/package.json
@@ -19,6 +19,7 @@
   },
   "exports": {
     ".": "./src/index.js",
+    "./zone-helpers.js": "./zone-helpers.js",
     "./heap.js": "./heap.js",
     "./tools/*": "./tools/*"
   },
@@ -27,6 +28,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@agoric/store": "^0.9.2",
+    "@endo/common": "^1.1.0",
     "@endo/exo": "^1.2.1",
     "@endo/far": "^1.0.4",
     "@endo/pass-style": "^1.2.0",
@@ -34,6 +36,7 @@
   },
   "devDependencies": {
     "@endo/init": "^1.0.4",
+    "@endo/ses-ava": "^1.1.2",
     "ava": "^5.3.0"
   },
   "publishConfig": {
