+++ packages/base-zone/src/prepare-revocable.js
